requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# LangChain Core
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-text-splitters==0.3.2

# LangGraph for Agentic Workflows
langgraph==0.2.45

# LangSmith for Tracing (Optional)
langsmith==0.1.140

# Vector Store & Embeddings
faiss-cpu==1.9.0
sentence-transformers==3.2.1

# PDF Processing
pypdf==5.1.0

# Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3

# Environment Variables
python-dotenv==1.0.1

# Utilities
typing-extensions==4.12.2
